change libscpi to current master & hack arb
this implements a hack into the libscpi library. The hack bypasses the
regular parsing of arbitrary data. Now only the header is parsed. The
length information from the is used by another function which directly
copies the data to the destination. This removes the limitation of the
libscpi that all data has to fit into the parse buffer (at least for arb
data)